requirements.txt
celery==5.2.7
Django==4.1.4
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1

